api/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
beanie==2.0.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
coverage==7.13.4
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.128.8
fastapi-cli==0.0.20
fastapi-cloud-cli==0.11.0
fastar==0.8.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
lazy-model==0.4.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mongomock==4.3.0
mongomock-motor==0.0.36
motor==3.7.1
packaging==26.0
passlib==1.7.4
pluggy==1.6.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pymongo==4.16.0
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
pytest-mock==3.15.1
python-dotenv==1.2.1
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rich==14.3.1
rich-toolkit==0.17.1
rignore==0.7.6
sentinels==1.1.1
sentry-sdk==2.50.0
setuptools==80.10.2
shellingham==1.5.4
starlette==0.52.1
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
wheel==0.46.3
